+++ scripts/executeScript/Set image size/Set image size - functions.jsx
@@ -1,0 +1,27 @@
+#include "../Ι_utils/functions/restrictInputKeys.jsx"
+
+#include "../Ι_utils/functions/leftUpperCornerColorBGSet.jsx";
+
+#include "../Ι_utils/functions/allowMinusOnlyAtFront.jsx";
+
+#include "../Ι_utils/functions/setMaxNumber.jsx";
+
+#include "../Ι_utils/functions/sameInputField.jsx";
+
+#include "../Ι_utils/functions/setDropdownSelectionFromEvent.jsx";
+
+#include "../Ι_utils/functions/changeTooltipAndImageAccordingToState.jsx";
+
+#include "../Ι_utils/functions/doesItHaveBackgroundLayer.jsx";
+
+#include "../Ι_utils/functions/mathSumWidthAndHeight.jsx";
+
+#include "../Ι_utils/functions/ErrorDiffrentUnitTypes.jsx";
+
+#include "../Ι_utils/functions/getRidOfTooMuch0AtFront.jsx";
+
+#include "../Ι_utils/functions/restrictValueUpTo.jsx";
+
+#include "../Ι_utils/functions/setMinimalValueAt.jsx";
+
+#include "../Ι_utils/functions/anchorSetingNew.jsx";