+++ scripts/main/main - EventHandlerBuilderMain.jsx
@@ -2,7 +2,7 @@
 using from "../control panel/control panel - EventHandlerBuilderControlPanel.jsx":
 - executeScript
 
-using from "../executeScript/((2^n canvas) || (Add canvas) || (Resize image))/((2^n canvas) || (Add canvas) || (Resize image)) - EventHandlerBuilderMain.jsx":
+using from "../executeScript/((2^n canvas) || (Add canvas) || (Add resize image))/((2^n canvas) || (Add canvas) || (Add resize image)) - EventHandlerBuilderMain.jsx":
 - self.lockingUnlockingAcceptBtn()
 
 using from "../settings/settings - functions.jsx":